пер. pageContent переименована в main, введена пер. layout, ф-я обрез-я текста перемещена в helpers.php
